extending hate speech api for twitter posts
