+++ lib/fetching/content-service.js
@@ -5,6 +5,9 @@
 var EventEmitter = require('events').EventEmitter;
 var util = require('util');
 var _ = require('underscore');
+var Promise = require('promise');
+var request = require('request');
+
 
 // options.lastReportDate - The time of the last report already fetched (optional)
 var ContentService = function(options) {
@@ -73,5 +76,35 @@ ContentService.prototype.fetch = function(options, callback) {
 ContentService.prototype._doFetch = function(options, callback) {
   throw new Error('Not yet implemented');
 };
+
+ContentService.prototype.getHateSpeechRequest = function(content) {
+	var self = this;
+	var options = { 
+		method: 'POST',
+  		url: 'http://localhost:5000/predict',
+		headers: { 
+			'Content-Type': 'application/json',
+			'Cache-Control': 'no-cache'
+		},
+  		body: {inputData: content },
+		json: true 
+  };
+	return new Promise(function(resolve, reject) {
+		request(options, function(err, res, body) {
+			if (err) {
+				self.emit('error', new Error('HTTP error: ' + err.message));
+				reject(err);
+			} else if (res.statusCode != 200) {
+				self.emit('error', new Error.HTTP(res.statusCode));
+				reject(err);
+			}
+			else{
+				resolve(body);
+			}
+
+		});
+	  });
+}
+
 
 module.exports = ContentService;
