+++ lib/fetching/content-services/crowd-tangle-content-service.js
@@ -78,7 +78,7 @@ CrowdTangleContentService.prototype._doFetch = function(options, callback) {
 			// the responses will be sorted by the content-service (parent method)
 			// Parse response data and return them.
 			if (self.detectHateSpeech){
-				var hateSpeechRequests = responses.map(function(x) {return {"data" : x, "request" :self.getHateSpeechRequest(x)}; });
+				var hateSpeechRequests = responses.map(function(x) {return {"data" : x, "request" :ContentService.prototype.getHateSpeechRequest(x.message ? x.message : "")}; });
 				Promise.all(hateSpeechRequests.map(function(x){return x.request})).then(
 					function(hateSpeechResults){
 						hateSpeechResults.forEach((hateSpeechResult, index)=>{
@@ -172,7 +172,7 @@ CrowdTangleContentService.prototype._parse = function(data, options) {
 		actualStatistics: data.statistics.actual || null,
 		expectedStatistics: data.statistics.expected || null,
 		rawAPIResponse: data,
-		hateSpeechScore: data.hateSpeechScore
+		hateSpeechScore: data.hateSpeechScore || null
 	};
 
 	var author = data.account ? data.account.name || data.account.handle : null;
@@ -217,36 +217,7 @@ CrowdTangleContentService.prototype._parse = function(data, options) {
   	};
 }
 
-CrowdTangleContentService.prototype.getHateSpeechRequest = function(data) {
-	var self = this;
-	var content = data.message ? data.message: "";
-	var options = { 
-		method: 'POST',
-  		url: 'http://localhost:5000/predict',
-		headers: { 
-			'Content-Type': 'application/json',
-			'Cache-Control': 'no-cache'
-		},
-  		body: {inputData: content },
-		json: true 
-	};
 
-	return new Promise(function(resolve, reject) {
-		request(options, function(err, res, body) {
-			if (err) {
-				self.emit('error', new Error('HTTP error: ' + err.message));
-				reject(err);
-			} else if (res.statusCode != 200) {
-				self.emit('error', new Error.HTTP(res.statusCode));
-				reject(err);
-			}
-			else{
-				resolve(body);
-			}
-
-		});
-	  });
-}
 
 CrowdTangleContentService.prototype.reloadSettings = function() {
 	this._baseUrl = config.get().crowdtangle.baseUrl;
