+++ lib/fetching/content-services/twitter-content-service.js
@@ -7,6 +7,7 @@ var config = require('../../../config/secrets');
 var util = require('util');
 var _ = require('lodash');
 var tweetMatches = require('tweet-matches');
+var Promise = require('promise');
 
 function TwitterContentService() {
   this._twit = new Twit(_.clone(config.get().twitter));
@@ -15,6 +16,7 @@ function TwitterContentService() {
   this.fetchType = 'subscribe';
   this._isStreaming = false;
   this._streamName = 'statuses/filter';
+  this.detectHateSpeech = config.get().crowdtangle.detectHateSpeech;
 }
 
 util.inherits(TwitterContentService, ContentService);
@@ -100,7 +102,24 @@ TwitterContentService.prototype._makeReport = function(tweet) {
   if (botIds.length === 0) {
     throw new Error('tweet-matches failed with keywords ' + this._keywords + ' on tweet http://twitter.com/' + (tweet.user && tweet.user.name) + '/status/' + tweet.id_str);
   }
-  this.emit('tweet:' + botIds[0], this._parse(tweet));
+  if (self.detectHateSpeech){
+    var hateSpeechRequests = [{"data" : tweet, "request" : ContentService.prototype.getHateSpeechRequest(tweet.text ? tweet.text : "")}];
+    Promise.all(hateSpeechRequests.map(function(x){return x.request})).then(
+      function(hateSpeechResults){
+        hateSpeechResults.forEach((hateSpeechResult, index)=>{
+          hateSpeechRequests[index].data.hateSpeechScore = hateSpeechResult.result.hateSpeechScore;
+        });
+        return hateSpeechRequests;
+      }
+    ).then((hateSpeechRequests) => {
+      //hateSpeechRequests will always have one element because Twitter gets tweets
+      //one at a time.
+      this.emit('tweet:' + botIds[0], this._parse(hateSpeechRequests[0].data));
+    });
+  } else{
+    this.emit('tweet:' + botIds[0], this._parse(tweet));
+  }
+  
 };
 
 // Gets a Twit stream object
@@ -123,6 +142,7 @@ TwitterContentService.prototype._parse = function(data) {
     longitude: data.coordinates ? data.coordinates[1] : 0,
     retweetCount: data.retweet_count ? data.retweet_count : 0,
     favouriteCount: data.favorite_count ? data.favorite_count : 0,
+    hateSpeechScore: data.hateSpeechScore || null,
     rawAPIResponse: data
   };
   return {
@@ -146,4 +166,4 @@ TwitterContentService.prototype.reloadSettings = function() {
   this._start();
 };
 
-module.exports = new TwitterContentService();
+module.exports = TwitterContentService;
